[update]#136 Changed putLike() and deleteLike() in PostRepository
* PostRepositoryInterfaceのputLike()とdeleteLike()の戻り値をnull許容に変更しました。
* EloquentPostRepositoryのputLike()とdeleteLike()で、対象の記事がない場合はnullを返却するように変更しました。

+++ laravel/app/Domain/Post/Repository/EloquentPostRepository.php
@@ -94,19 +94,26 @@ class EloquentPostRepository implements PostRepository
         return $posts;
     }
 
-    public function putLike(PostId $postId, UserAccountId $userId): PostId
+    public function putLike(PostId $postId, UserAccountId $userId)
     {
         $post = Post::where('id', $postId->toString())->with(['likes'])->first();
 
+        if (!$post) {
+            return;
+        }
         $post->likes()->detach($userId->toInt());
         $post->likes()->attach($userId->toInt());
 
         return $postId;
     }
 
-    public function deleteLike(PostId $postId, UserAccountId $userId): PostId
+    public function deleteLike(PostId $postId, UserAccountId $userId)
     {
         $post = Post::where('id', $postId->toString())->with(['likes'])->first();
+
+        if (!$post) {
+            return;
+        }
 
         $post->likes()->detach($userId->toInt());
